Update package versions and switch to async app initialization
The commit updates the Aiursoft.WebTools package to version 8.0.12 in both SampleWebApp and AiurEventSyncer.WebExtends projects. AiurEventSyncer.WebExtends project version is also bumped up to 8.0.4. The Main method in Program.cs and the CreateServer method in BasicTests.cs are updated to call the asynchronous version of the app initialization function.